refactor: newsletter preview component
